cloud functions/otp trigger/requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-tasks
google-cloud-core
twilio
requests
